pyproject.toml
[tool.poetry]
name = "connect-markdown-renderer"
version = "0.0.0"
description = "Connect Markdown Renderer"
authors = ["CloudBlue"]
license = "Apache-2.0"
packages = [
    { include = "connect" },
    { include = "cmr" }
]
readme = "./README.md"
documentation = "https://github.com/cloudblue/connect-markdown-renderer"
homepage = "https://connect.cloudblue.com"
repository = "https://github.com/cloudblue/connect-markdown-renderer"

[tool.poetry.dependencies]
python = ">=3.7,<4"
rich = "^12.4.1"
markdown-it-py = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
pytest-cov = "^2.10.1"
pytest-mock = "^3.3.1"
coverage = {extras = ["toml"], version = "^5.3"}
flake8 = "~3.8"
flake8-bugbear = "~20"
flake8-cognitive-complexity = "^0.1"
flake8-commas = "~2.0"
flake8-future-import = "~0.4"
flake8-import-order = "~0.18"
flake8-broken-line = "~0.3"
Sphinx = "^4.5.0"
sphinx-copybutton = "^0.5.0"
m2r2 = "^0.3.2"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=42"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = "tests"
addopts = "--cov=connect.utils.terminal --cov-report=term-missing:skip-covered --cov-report=html --cov-report=xml"

[tool.coverage.run]
branch = true

[tool.coverage.report]
omit = [
]

exclude_lines = [
    "pragma: no cover",
    "def __str__",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]